+++ nac/basisSet/__init__.py
@@ -1,6 +1,0 @@
-from .basisNormalization import (
-    compute_normalization_sphericals, create_dict_CGFs, create_normalized_CGFs)
-from .contractedGFs import createUniqueCGF
-
-__all__ = ['compute_normalization_sphericals', 'create_dict_CGFs',
-           'create_normalized_CGFs', 'createUniqueCGF']