+++ nac/integrals/__init__.py
@@ -1,15 +1,7 @@
-from .multipole_matrices import (compute_matrix_multipole, get_multipole_matrix)
-from .multipoleIntegrals import (calcMtxMultipoleP, general_multipole_matrix)
-from .nonAdiabaticCoupling import (calculate_couplings_levine,
-                                   calculate_couplings_3points,
-                                   compute_overlaps_for_coupling,
-                                   correct_phases)
-from .overlapIntegral import calcMtxOverlapP
-from .spherical_Cartesian_cgf import calc_transf_matrix
-
+from .nonAdiabaticCoupling import (
+    calculate_couplings_levine, calculate_couplings_3points,
+    compute_overlaps_for_coupling, correct_phases)
 
 __all__ = ['calculate_couplings_3points', 'calculate_couplings_levine',
-           'calc_transf_matrix', 'calcMtxMultipoleP', 'calcMtxOverlapP',
-           'calculate_couplings_levine', 'compute_matrix_multipole',
-           'compute_overlaps_for_coupling', 'correct_phases',
-           'general_multipole_matrix', 'get_multipole_matrix']
+           'calculate_couplings_levine', 'compute_overlaps_for_coupling',
+           'correct_phases']